add Licence file with information from readme.html
This fork based on verion 1.7.4 where monroewilliams allready removed Sparkle.framework and QuickTime.framework from the linked frameworks in the project.

- just open project in Xcode 10.3 (no automated import changes are done jet !)
- add licence information in an seperate file because licence information was kind of hidden at end of resource file 'English.lproj/readme.html'
- remove sparcle.framework resources completly (was old 1.1 version) since it is not longer used since changes from monroewilliams

There are still 62 warnings that seems to needed more work to fix. But ignoring those it compiles anyway to a 64-bit application that runs fine so far.

Since the original webserver that delivered this free source code is not longer on the web here is the link to the webarchive showing the original readmefile:
https://web.archive.org/web/20161018151125/http://dozingcatsoftware.com/SaverLab/readme.html